Added iframe-game
This is the side project I've mentioned before.  There are two html files and a css file in it, but at the moment none of it really has anything in it.  More will come later, I just wanted to set up the structure now so I don't keep putting it off.